Enemy Pathing
Some progress was made towards getting the enemy pathing script to work
the way I want it. I made a text document describing all of the changes
I want to make, and I am making changes based on that. Currently
PathFinder is structured exactly how I need it to be, which basically
means it's just used to generate the path, and AIMovement is going to be
used to traverse the path and determine whether or not a new path is
necessary or not.