Initially Finished AI Pathing
AI Pathing's code is finished between PathFinder and AIMovement,but has
not been tested yet. Tweaks can be made if there are issues using it
with the enemy. Additionally, all prefab objects related to the old
tank, and the "non-modified" new tank have been deleted, so-as to reduce
future confusion.