Working AI Pathing
Most of the pieces of AI pathing have been implemented and are working.
Some additional things such as determining when the enemies are going to
explode, and whether or not those explosions are going to damage the
resource nodes still need to be done. Most of it is now working,
however. Every 20 seconds 16 enemies spawn in the center of the map and
chase a random target.