Terrian wall creation and destruction
- wall prefab description
The wall prefab is just a cube some scripts attached to it
Attached scripts - wallScript
Location - Base folder in prefab folder

- wallScript Description
Attached to the wall
The wallScript pretty much creates a health for the and after a certian
number of hits the wall instantiates another new prefab walldetonation
for a cool explosion effect.
Attached prefabs - wallDetonation prefab
Location - the AI script folder in the script folders

- wallDetonation prefab
this is a custom detonation prefab that was created for when the wall
blows up. makes use of the detonationSpray script from the detonation
framework to create an explosion with multiple chunk prefabs to fly up
in the air
Attached - chunk prefab
Location - Tank folder under prefab folder

- detonatorForce script modifications
The detonatorforce script in the detonator framework has also been
modified to check for explosions near the wall tag.