Finished AI Pathing
AI Pathing now covers the issue of not being able to find the transform
of the object. If it no longer exists, the enemies will look for a new
one. If there are no more new targets, they will automatically explode.
Unfortunately, if there are targets but nothing can see them Unity will
generate 1000 errors per second again. That is a more complicated error
to fix.

I also got rid of lots of whitespace and changed the terrain and some of
the prefabs so that they would work with the new AI pathing.