Retargetting When Node Dies
Enemies previously were not selecting new targets when a node died. This
was because of a change in tags that made it so that enemies could not
explode when reaching a crystal resource rather than a box resource. In
addition, apparently RPC's can't send a network view but can send a
network view ID across the network.