Fixed AudioListener Complaints
Previously the Unity engine was complaining that there was not exactly
one audio listener in the scene. It would either be because there were
none or more than one. Now, there is a new camera which will prevent
this issue from occurring on the server side, which is disabled if the
ClientPlayerController is on the client side. In addition, the
server-side code now disables all of the tank's listeners.