Implemented Mortar Weapon and AIMovement Fix
The mortar now works. It charges for a maximum of 1.5 seconds for a
total power boost of 45. Things like the weight of the mortar shot and
the max launching speed of the mortar can be adjusted later if launching
really far becomes a problem.

In addition, I added a couple return; statements in AIMovement to
hopefully once and for all prevent the code from triggering when it
self-destructs.