Your Blip Appears On Enemy Radar
The model has been messed around with in several ways to make this work,
but it works now. Your blip also appears on your radar, but this can be
changed later. In addition, the cooldown of the radar has been reduced
to 5 seconds to make it easier to test with.

In addition, the enemy generator now has a toggle that prevents enemies
from spawning if you don't want them to.