Machine Gun, Enemy Gen, and Movement Changes
I have discovered the reason why we couldn't damage using machine gun
before, and that is because all of our objects were previously using the
ignore raycast layer. I changed this, and the pathfinding, so that it
would not cause issues to not use the ignore raycast layer.

In addition, the forward movement rotation now has a check on it to make
it less difficult to move around.

The last change I made was to make it so that the number of enemies
spawning could be changed to whatever we wanted, and I made it so that
currently enemies spawning are just immobile targets. This can be
changed in-game still to get enemies that come after you.