Added Lightmapping to Important Scenes
I was looking into ways to increase the performance of our game, and I
completely forgot about baking a lightmap of our scene so that we could
take advantage of static objects.

I did this for the rocky crag, and every scene with even a little bit of
textured stuff on it. This should drastically improve the performance of
our game. I'm looking into other changes we could make.