Started on Lobby and Nodes Damageable
The lobby has most of the components in it that will make it work, but I
still need to test it to make sure it works and to change the splash
controller to make it go to the lobby instead of directly to rocky crag.

Also applied a quick fix that makes it so nodes can be damaged by the
mortar bullet, any kind of detonator framework explosion, and the
machine gun raycast.