Added Volume Control GUI
I found a little javascript that I converted to C#, which allows you to
control an audio source's volume (and possibly other things) with a
slider. I set the slider up to show its volume on a 1-100 scale.

In addition, I disabled a few things on the server side which should
reduce the amount of graphics processing on the server and make
calculations less of an issue.

Finally, I made it so that the player views good graphics on the scene.
It was a camera setting. Apparently if forward rendering is not enabled
(at least) then it has really bad resolution (it has to be enabled on
the active camera in the scene).