Movement and TurretRotation Changes
I finished assuring that movement and tank turret rotation would work
with the new model. I also got rid of the old movement, turret rotation,
and shooting scripts, because they are going to be replaced with the new
controllers. These will support both the RPC method of control and
client-side method of control.