Mortar and Machine Gun Changes
I added new scripts for the shooting controls which we should be able to
use and the same modifications for those scripts on the server player
controller.

I was attempting to break them down into more methods like I did for the
movement and turret rotation scripts, but actually there are only two
code paths for each anyway, so I couldn't find any way to simplify
either of them at a glance. This might change later.