Client Side Prediction and Weird Fixes
Hooray! We have client-side movement prediction! This basically means
the clients will help in predicting how they will move, and so networked
movement will be a bit smoother (it's not going to be perfect).

In addition I got rid of or commented out several print statements that
happened a lot.

I also figured out why we couldn't move the tank, and it's weird, but
just don't mess with the audio listener code fixes anymore.

The last thing I did was halve the rotation speed of the tank. I thought
it was a little too fast and it seemed to help the smoothness of the
rotation a bit.