Client Side Prediction Improvements
The lerping function that I was using in order to update the client's
tank wasn't quite as smooth as I would have liked and behaved weirdly,
so I decided to tweak it. And I tweaked it until the rotation was
perfectly smooth and the movement was only a little bit behind the
server movement and also still very smooth.