Nodes Damageable Rigidbody Issue
The detonator framework code had to be changed again in order to make
sure that the code for damaging the nodes, and a rigidbody had to be
added to each of the nodes in the scene, but it now works for mortars
too.

In addition, the machine gun bullets were not being destroyed on the
client side (hard to notice before, without the unity client looking at
game object references), and so a new script has been added to the
machine gun bullet that destroys them upon collision over the network.