Spawn Color Order Fixed
Players will now spawn in the correct order, with the correct colors, in
the lobby and the rocky crag scene, and anywhere else we decide to put
them. Apparently there is no way to guarantee that we will get the same
spawn order unless we hard-code a name look-up at the start, and that is
what I did.